Added 'pack' and 'unpack' methods to CwtUtilsNS. Created appropriate tests.

+++ src/core/fsm.c
@@ -183,7 +183,7 @@ static ssize_t __fsm_exec(CwtFsm *fsm, int state_cur, const void *data, size_t l
 		switch( trans->match_type ) {
 		case Cwt_Fsm_Match_Seq:
 			if( len >= trans->condition.str.len )
-				nchars_processed = trans->condition.str.len;
+				nchars_processed = (ssize_t)trans->condition.str.len;
 			break;
 
 		case Cwt_Fsm_Match_Str:
